requirements.txt
Django
DjangoRestFramework